refactor: extract shared AgentContentTransformer from AgentHistoryEntityTransformer
Move the content-block mapping logic (text/document-reference blocks)
out of AgentHistoryEntityTransformer into a standalone
AgentContentTransformer so it can be reused for the agent instance
system prompt. Behavior is unchanged.
