+++ search/search-client-query-transformer/src/main/java/io/camunda/search/clients/transformers/entity/AgentHistoryEntityTransformer.java
@@ -15,16 +15,10 @@ import io.camunda.search.entities.AgentInstanceHistoryEntity.Limits;
 import io.camunda.search.entities.AgentInstanceHistoryEntity.Metrics;
 import io.camunda.search.entities.AgentInstanceHistoryEntity.Tool;
 import io.camunda.search.entities.AgentInstanceHistoryEntity.ToolCall;
-import io.camunda.search.entities.ContentItem;
-import io.camunda.search.entities.ContentItem.ContentType;
-import io.camunda.search.entities.DocumentMetadata;
-import io.camunda.search.entities.DocumentReference;
-import io.camunda.webapps.schema.entities.agenthistory.AgentHistoryContentValue;
 import io.camunda.webapps.schema.entities.agenthistory.AgentHistoryEntity;
 import io.camunda.webapps.schema.entities.agenthistory.AgentHistoryEntity.AgentHistoryEmbeddedToolCallValue;
 import io.camunda.webapps.schema.entities.agenthistory.AgentHistoryEntity.AgentHistoryLimitsValue;
 import io.camunda.webapps.schema.entities.agenthistory.AgentHistoryEntity.AgentHistoryToolValue;
-import io.camunda.webapps.schema.entities.document.DocumentReferenceEntity;
 import java.util.List;
 import java.util.Objects;
 
@@ -46,14 +40,14 @@ public class AgentHistoryEntityTransformer
         source.getJobLease(),
         source.getLoopIteration(),
         toRole(source.getRole()),
-        toContent(source.getContent()),
+        AgentContentTransformer.toContent(source.getContent()),
         toToolCalls(source.getToolCalls()),
         toMetrics(source.getInputTokens(), source.getOutputTokens(), source.getDurationMs()),
         toTools(source.getTools()),
         source.getModel(),
         source.getProvider(),
         toLimits(source.getLimits()),
-        toSystemPrompt(source.getSystemPrompt()),
+        AgentContentTransformer.toContent(source.getSystemPrompt()),
         toCommitStatus(source.getCommitStatus()),
         source.getProducedAt());
   }
@@ -75,42 +69,6 @@ public class AgentHistoryEntityTransformer
       case COMMITTED -> AgentInstanceHistoryCommitStatus.COMMITTED;
       case DISCARDED -> AgentInstanceHistoryCommitStatus.DISCARDED;
     };
-  }
-
-  private static List<ContentItem> toContent(final List<AgentHistoryContentValue> content) {
-    if (content == null) {
-      return List.of();
-    }
-    return content.stream().map(AgentHistoryEntityTransformer::toContentItem).toList();
-  }
-
-  private static ContentItem toContentItem(final AgentHistoryContentValue value) {
-    return switch (value.contentType()) {
-      case TEXT -> new ContentItem(ContentType.TEXT, value.text(), null, null);
-      case DOCUMENT ->
-          new ContentItem(
-              ContentType.DOCUMENT, null, toDocumentReference(value.documentReference()), null);
-      case OBJECT -> new ContentItem(ContentType.OBJECT, null, null, value.object());
-    };
-  }
-
-  private static DocumentReference toDocumentReference(final DocumentReferenceEntity entity) {
-    if (entity == null) {
-      return null;
-    }
-    final var meta = entity.metadata();
-    return new DocumentReference(
-        entity.storeId(),
-        entity.documentId(),
-        entity.contentHash(),
-        new DocumentMetadata(
-            meta.contentType(),
-            meta.fileName(),
-            meta.expiresAt(),
-            meta.size(),
-            meta.processDefinitionId(),
-            meta.processInstanceKey(),
-            meta.customProperties()));
   }
 
   private static List<ToolCall> toToolCalls(
@@ -153,14 +111,5 @@ public class AgentHistoryEntityTransformer
       return new Limits(-1, -1, -1);
     }
     return new Limits(limits.maxTokens(), limits.maxModelCalls(), limits.maxToolCalls());
-  }
-
-  /** The system prompt, as content blocks, as of this entry. CONFIGURATION items only. */
-  private static List<ContentItem> toSystemPrompt(
-      final List<AgentHistoryContentValue> systemPrompt) {
-    if (systemPrompt == null) {
-      return List.of();
-    }
-    return systemPrompt.stream().map(AgentHistoryEntityTransformer::toContentItem).toList();
   }
 }
